TransferEntropy.jl

This package exports two functions, transferentropy and mutualinfo. Both functions use the estimators listed below (not all estimators are implemented for both functions; see docstrings for transferentropy and mutualinfo for details).

Estimators

Binning based

Entropies.VisitationFrequency — Type

VisitationFrequency(r::RectangularBinning) <: BinningProbabilitiesEstimator

A probability estimator based on binning data into rectangular boxes dictated by the binning scheme r.

Example

# Construct boxes by dividing each coordinate axis into 5 equal-length chunks.
b = RectangularBinning(5)

# A probabilities estimator that, when applied a dataset, computes visitation frequencies
# over the boxes of the binning, constructed as describedon the previous line.
est = VisitationFrequency(b)

See also: RectangularBinning.

Entropies.RectangularBinning — Type

RectangularBinning(ϵ) <: RectangularBinningScheme

Instructions for creating a rectangular box partition using the binning scheme ϵ. Binning instructions are deduced from the type of ϵ.

Rectangular binnings may be automatically adjusted to the data in which the RectangularBinning is applied, as follows:

ϵ::Int divides each coordinate axis into ϵ equal-length intervals, extending the upper bound 1/100th of a bin size to ensure all points are covered.

ϵ::Float64 divides each coordinate axis into intervals of fixed size ϵ, starting from the axis minima until the data is completely covered by boxes.

ϵ::Vector{Int} divides the i-th coordinate axis into ϵ[i] equal-length intervals, extending the upper bound 1/100th of a bin size to ensure all points are covered.

ϵ::Vector{Float64} divides the i-th coordinate axis into intervals of fixed size ϵ[i], starting from the axis minima until the data is completely covered by boxes.

Rectangular binnings may also be specified on arbitrary min-max ranges.

ϵ::Tuple{Vector{Tuple{Float64,Float64}},Int64} creates intervals along each coordinate axis from ranges indicated by a vector of (min, max) tuples, then divides each coordinate axis into an integer number of equal-length intervals. Note: this does not ensure that all points are covered by the data (points outside the binning are ignored).

Example 1: Grid deduced automatically from data (partition guaranteed to cover data points)

Flexible box sizes

The following binning specification finds the minima/maxima along each coordinate axis, then split each of those data ranges (with some tiny padding on the edges) into 10 equal-length intervals. This gives (hyper-)rectangular boxes, and works for data of any dimension.

using Entropies
RectangularBinning(10)

Now, assume the data consists of 2-dimensional points, and that we want a finer grid along one of the dimensions than over the other dimension.

The following binning specification finds the minima/maxima along each coordinate axis, then splits the range along the first coordinate axis (with some tiny padding on the edges) into 10 equal-length intervals, and the range along the second coordinate axis (with some tiny padding on the edges) into 5 equal-length intervals. This gives (hyper-)rectangular boxes.

using Entropies
RectangularBinning([10, 5])

Fixed box sizes

The following binning specification finds the minima/maxima along each coordinate axis, then split the axis ranges into equal-length intervals of fixed size 0.5 until the all data points are covered by boxes. This approach yields (hyper-)cubic boxes, and works for data of any dimension.

using Entropies
RectangularBinning(0.5)

Again, assume the data consists of 2-dimensional points, and that we want a finer grid along one of the dimensions than over the other dimension.

The following binning specification finds the minima/maxima along each coordinate axis, then splits the range along the first coordinate axis into equal-length intervals of size 0.3, and the range along the second axis into equal-length intervals of size 0.1 (in both cases, making sure the data are completely covered by the boxes). This approach gives a (hyper-)rectangular boxes.

using Entropies
RectangularBinning([0.3, 0.1])

Example 2: Custom grids (partition not guaranteed to cover data points):

Assume the data consists of 3-dimensional points (x, y, z), and that we want a grid that is fixed over the intervals [x₁, x₂] for the first dimension, over [y₁, y₂] for the second dimension, and over [z₁, z₂] for the third dimension. We when want to split each of those ranges into 4 equal-length pieces. Beware: some points may fall outside the partition if the intervals are not chosen properly (these points are simply discarded).

The following binning specification produces the desired (hyper-)rectangular boxes.

using Entropies, DelayEmbeddings

D = Dataset(rand(100, 3));

x₁, x₂ = 0.5, 1 # not completely covering the data, which are on [0, 1]
y₁, y₂ = -2, 1.5 # covering the data, which are on [0, 1]
z₁, z₂ = 0, 0.5 # not completely covering the data, which are on [0, 1]

ϵ = [(x₁, x₂), (y₁, y₂), (z₁, z₂)], 4 # [interval 1, interval 2, ...], n_subdivisions

RectangularBinning(ϵ)

Kernel density based

Entropies.NaiveKernel — Type

NaiveKernel(ϵ::Real, method::KernelEstimationMethod = TreeDistance()) <: ProbabilitiesEstimator

Estimate probabilities/entropy using a "naive" kernel density estimation approach (KDE), as discussed in Prichard and Theiler (1995) [PrichardTheiler1995].

Probabilities $P(\mathbf{x}, \epsilon)$ are assigned to every point $\mathbf{x}$ by counting how many other points occupy the space spanned by a hypersphere of radius ϵ around $\mathbf{x}$, according to:

\[P_i( \mathbf{x}, \epsilon) \approx \dfrac{1}{N} \sum_{s \neq i } K\left( \dfrac{||\mathbf{x}_i - \mathbf{x}_s ||}{\epsilon} \right),\]

where $K(z) = 1$ if $z < 1$ and zero otherwise. Probabilities are then normalized.

Methods

Tree-based evaluation of distances using TreeDistance. Faster, but more memory allocation.

Direct evaluation of distances using DirectDistance. Slower, but less memory allocation. Also works for complex numbers.

Estimation

Probabilities or entropies can be estimated from Datasets.

probabilities(x::AbstractDataset, est::NaiveKernel). Associates a probability p to each point in x.

genentropy(x::AbstractDataset, est::NaiveKernel). Associate probability p to each point in x, then compute the generalized entropy from those probabilities.

Examples

using Entropy, DelayEmbeddings
pts = Dataset([rand(5) for i = 1:10000]);
ϵ = 0.2
est_direct = NaiveKernel(ϵ, DirectDistance())
est_tree = NaiveKernel(ϵ, TreeDistance())

p_direct = probabilities(pts, est_direct)
p_tree = probabilities(pts, est_tree)

# Check that both methods give the same probabilities
all(p_direct .== p_tree)

See also: DirectDistance, TreeDistance.

Entropies.TreeDistance — Type

TreeDistance(metric::M = Euclidean()) <: KernelEstimationMethod

Pairwise distances are evaluated using a tree-based approach with the provided metric.

Entropies.DirectDistance — Type

DirectDistance(metric::M = Euclidean()) <: KernelEstimationMethod

Pairwise distances are evaluated directly using the provided metric.

Nearest neighbor based

Entropies.KozachenkoLeonenko — Type

Nearest neighbour(NN) based

KozachenkoLeonenko(; w::Int = 0) <: NearestNeighborEntropyEstimator

Entropy estimator based on nearest neighbors. This implementation is based on Kozachenko & Leonenko (1987)[KozachenkoLeonenko1987], as described in Charzyńska and Gambin (2016)[Charzyńska2016].

w is the Theiler window (defaults to 0, meaning that only the point itself is excluded when searching for neighbours).

Info

This estimator is only available for entropy estimation. Probabilities cannot be obtained directly.

Entropies.Kraskov — Type

k-th nearest neighbour(kNN) based

Kraskov(k::Int = 1, w::Int = 1) <: NearestNeighborEntropyEstimator

Entropy estimator based on k-th nearest neighbor searches[Kraskov2004].

w is the number of nearest neighbors to exclude when searching for neighbours (defaults to 0, meaning that only the point itself is excluded).

Info

This estimator is only available for entropy estimation. Probabilities cannot be obtained directly.

TransferEntropy.Kraskov1 — Type

Kraskov1(k::Int = 1; metric_x = Chebyshev(), metric_y = Chebyshev()) <: MutualInformationEstimator

The $I^{(1)}$ nearest neighbor based mutual information estimator from Kraskov et al. (2004), using k nearest neighbors. The distance metric for the marginals $x$ and $y$ can be chosen separately, while the Chebyshev metric is always used for the z = (x, y) joint space.

TransferEntropy.Kraskov2 — Type

Kraskov2(k::Int = 1; metric_x = Chebyshev(), metric_y = Chebyshev()) <: MutualInformationEstimator

The $I^{(2)}(x, y)$ nearest neighbor based mutual information estimator from Kraskov et al. (2004), using k nearest neighbors. The distance metric for the marginals $x$ and $y$ can be chosen separately, while the Chebyshev metric is always used for the z = (x, y) joint space.

Transfer entropy

TransferEntropy.transferentropy — Function

General interface

transferentropy(s, t, [c], est; base = 2, α = 1, 
    τT = -1, τS = -1, η𝒯 = 1, dT = 1, dS = 1, d𝒯 = 1, [τC = -1, dC = 1])

Estimate transfer entropy from source s to target t, potentially conditioned on c ($TE(s \to t$/$TE(s \to t | c)$),with logarithms to the given base, using the provided entropy/probability estimator est.

The input series s, t, and c are equal-length real-valued vectors of length N.

The relation between the embedding lags τT, τS, τC, the η𝒯 (prediction lag), and the embedding dimensions dT, dS, dC, d𝒯 is given below.

Nearest neighbor based

transferentropy(s, t, [c], est::Kraskov; base = 2, ...)
transferentropy(s, t, [c], est::KozachenkoLeonenko; base = 2, ...)

Estimate $TE(s \to t$/$TE(s \to t | c)$ using naive nearest neighbor estimators.

For these estimators, only Shannon entropy can be computed (so the keyword α does not work).

See also Kraskov, KozacheckoLeonenko.

Kernel density based

transferentropy(s, t, [c], est::NaiveKernel{Union{TreeDistance, DirectDistance}}; 
    base = 2, α = 1,  ...)

Estimate $TE(s \to t$/$TE(s \to t | c)$ using naive kernel density estimation of probabilities.

See also NaiveKernel, TreeDistance, DirectDistance.

Instantenous Hilbert amplitudes/phases

transferentropy(s, t, [c], est::Hilbert; base = 2, α = 1,  ...)

Estimate $TE(s \to t$/$TE(s \to t | c)$ by first applying the Hilbert transform to s, t (c) and then estimating transfer entropy.

See also [Hilbert], Amplitude, Phase.

Symbolic/permutation

transferentropy(s, t, [c], est::SymbolicPermutation; 
    base = 2, α = 1, m::Int = 3, τ::Int = 1, ...)
transferentropy!(symb_s, symb_t, s, t, [c], est::SymbolicPermutation; 
    base = 2, α = 1, m::Int = 3, τ::Int = 1, ...)

Estimate $TE(s \to t$/$TE(s \to t | c)$ using permutation entropies. This is done by first symbolizing the input series s and t (both of length N) using motifs of size m and a time delay of τ. The series of motifs are encoded as integer symbol time series preserving the permutation information. These symbol time series are embedded as usual, and transfer entropy is computed from marginal entropies of that generalized embedding.

Optionally, provide pre-allocated (integer) symbol vectors symb_s and symb_t, where length(symb_s) == length(symb_t) == N - (est.m-1)*est.τ. This is useful for saving memory allocations for repeated computations.

See also SymbolicPermutation.

Details

Transfer entropy (TE) between two simultaneously measured time series, $s(n) = \{ s_1, s_2, \ldots, s_N \}$ and $t(n) = \{ s_1, s_2, \ldots, s_N \}$, is defined as

\[TE(s \to t) = \sum_i p(s_n, t_n, t_{n+\eta}) \log \left( \dfrac{p(t_{n+\eta} | t_n, s_n)}{p(t_{n+\eta} | t_n)} \right)\]

Above, the TE is expressed in in simplest, scalar form. Including more than one historical/future value can be done by defining the generalized delay embeddings

$\mathcal{T}^{(d_{\mathcal T}, \eta_{\mathcal T})} = \{t_i^{(d_{\mathcal T}, \eta_{\mathcal T})} \}_{i=1}^{N}$

$T^{d_T, \tau_T} = \{t_i^{(d_T, \tau_T)} \}_{i=1}^{N}$

$S^{d_S, \tau_S} = \{s_i^{(d_T, \tau_T)} \}_{i=1}^{N}$,

$C^{d_C, \tau_C} = \{s_i^{(d_C, \tau_C)} \}_{i=1}^{N}$,

each having N distinct states, where the $d_T$-dimensional, $d_S$-dimensional and $d_C$-dimensional state vectors comprising $T$, $S$ and $C$ are constructed with embedding lags $\tau_T$, $\tau_S$, and $\tau_C$, respectively. The $d_{\mathcal T}$-dimensional future states $\mathcal{T}^{(d_{\mathcal T}, \eta_{\mathcal T})}$ are constructed with prediction lag $\eta_{\mathcal T}$ (i.e. predictions go from present/past states to future states spanning a maximum of $d_{\mathcal T} \eta_{\mathcal T}$ time steps ).

Note: in the original transfer entropy paper, only the historical states are defined as potentially higher-dimensional, while the future states are always scalar.

The non-conditioned and conditioned generalized forms of the transfer entropy is then

\[TE(s \to t) = \sum_i p(S,T, \mathcal{T}) \log \left( \dfrac{p(\mathcal{T} | T, S)}{p(\mathcal{T} | T)} \right)\]

\[TE(s \to t | c) = \sum_i p(S,T, \mathcal{T}, C) \log \left( \dfrac{p(\mathcal{T} | T, S, C)}{p(\mathcal{T} | T, C)} \right)\]

Estimation

Transfer entropy is here estimated by rewriting the generalized transfer entropy, using properties of logarithms and conditional probabilities, as a sum of marginal entropies

\[TE(s \to t) = H(\mathcal T, T) + H(\mathcal T, S) - H(T) - H(\mathcal T, T, S),\]

\[TE(s \to t | c) = H(\mathcal T, T, C) + H(\mathcal T, S, C) - H(T, C) - H(\mathcal T, T, S, C),\]

where $H(\cdot)$ is the generalized Renyi entropy. Individual marginal entropies are here computed using the provided estimator est. In the original transfer entropy paper, the Shannon entropy is used. Here, by adjusting the keyword α (defaults to α=1 for Shannon entropy), the transfer entropy, using the generalized Renyi enropy of order α, can be computed.

Mutual information

TransferEntropy.mutualinfo — Function

General interface

mutualinfo(x, y, est; base = 2, α = 1)

Estimate mutual information ($I$) between x and y using the provided entropy/probability estimator est, with logarithms to the given base. Optionally, use the generalized Rényi entropy of order α (defaults to α = 1, which is the Shannon entropy). See details below.

Both x and y can be vectors or (potentially multivariate) Datasets.

Binning based

mutualinfo(x, y, est::VisitationFrequency{RectangularBinning}; base = 2, α = 1)

Estimate $I(x, y)$ using a visitation frequency estimator.

See also VisitationFrequency, RectangularBinning.

Kernel density based

mutualinfo(x, y, est::NaiveKernel{Union{DirectDistance, TreeDistance}}; base = 2, α = 1)

Estimate $I(x, y)$ using a naive kernel density estimator.

It is possible to use both direct evaluation of distances, and a tree-based approach. Which approach is faster depends on the application.

See also NaiveKernel, DirectDistance, TreeDistance.

Nearest neighbor based

mutualinfo(x, y, est::KozachenkoLeonenko; base = 2)
mutualinfo(x, y, est::Kraskov; base = 2)
mutualinfo(x, y, est::Kraskov1; base = 2)
mutualinfo(x, y, est::Kraskov2; base = 2)

Estimate $I(x, y)$ using a nearest neighbor based estimator. Choose between naive estimation using the KozachenkoLeonenko or Kraskov entropy estimators, or the improved Kraskov1 and Kraskov2 dedicated $I$ estimators. The latter estimators reduce bias compared to the naive estimators.

Note: only Shannon entropy is possible to use for these estimators.

See also KozachenkoLeonenko, Kraskov, Kraskov1, Kraskov2.

Details/estimation

Mutual information is defined as

\[I(X; Y) = \sum_{y \in Y} \sum_{x \in X} p(x, y) \log \left( \dfrac{p(x, y)}{p(x)p(y)} \right)\]

This expression can be expressed as the sum of marginal entropies as follows:

\[I(X; Y) = H(X) + H(Y) - H(X, Y).\]

These individual entropies are computed using the provided entropy/probabilities estimator. For some estimators, it is possible to use generalized order-α Rényi entropies for the $I(x, y)$ computation, but the default is to use the Shannon entropy (α = 1).

Dataset

DelayEmbeddings.Dataset — Type

Dataset{D, T} <: AbstractDataset{D,T}

A dedicated interface for datasets. It contains equally-sized datapoints of length D, represented by SVector{D, T}. These data are contained in the field .data of a dataset, as a standard Julia Vector{SVector}.

When indexed with 1 index, a dataset is like a vector of datapoints. When indexed with 2 indices it behaves like a matrix that has each of the columns be the timeseries of each of the variables.

Dataset also supports most sensible operations like append!, push!, hcat, eachrow, among others.

Description of indexing

In the following let i, j be integers, typeof(data) <: AbstractDataset and v1, v2 be <: AbstractVector{Int} (v1, v2 could also be ranges).

data[i] gives the ith datapoint (returns an SVector)

data[v1] will return a vector of datapoints

data[v1, :] using a Colon as a second index will return a Dataset of these points

data[:, j] gives the jth variable timeseries, as Vector

data[v1, v2] returns a Dataset with the appropriate entries (first indices being "time"/point index, while second being variables)

data[i, j] value of the jth variable, at the ith timepoint

Use Matrix(dataset) or Dataset(matrix) to convert. It is assumed that each column of the matrix is one variable. If you have various timeseries vectors x, y, z, ... pass them like Dataset(x, y, z, ...). You can use columns(dataset) to obtain the reverse, i.e. all columns of the dataset in a tuple.

PrichardTheiler1995Prichard, D., & Theiler, J. (1995). Generalized redundancies for time series analysis. Physica D: Nonlinear Phenomena, 84(3-4), 476-493.

Charzyńska2016Charzyńska, A., & Gambin, A. (2016). Improvement of the k-NN entropy estimator with applications in systems biology. Entropy, 18(1), 13.

KozachenkoLeonenko1987Kozachenko, L. F., & Leonenko, N. N. (1987). Sample estimate of the entropy of a random vector. Problemy Peredachi Informatsii, 23(2), 9-16.

Kraskov2004Kraskov, A., Stögbauer, H., & Grassberger, P. (2004). Estimating mutual information. Physical review E, 69(6), 066138.